need to work on rating modal
